Move tests to commands
