+++ src/main/java/frc/robot/subsystems/drivetrain/commands/tests/AngleMotorsSpeedTest.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.drivetrain.tests;
+package frc.robot.subsystems.drivetrain.commands.tests;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
@@ -7,12 +7,15 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.drivetrain.SwerveDrive;
 
-public class DriveMotorsSpeedTest extends CommandBase {
+import java.util.Arrays;
+
+public class AngleMotorsSpeedTest extends CommandBase {
     private final SwerveDrive swerveDrive;
     private final Timer timer = new Timer();
     private final double[] times = new double[4];
+    private final SwerveModuleState[] states = new SwerveModuleState[4];
 
-    public DriveMotorsSpeedTest(SwerveDrive swerveDrive) {
+    public AngleMotorsSpeedTest(SwerveDrive swerveDrive) {
         this.swerveDrive = swerveDrive;
         addRequirements(swerveDrive);
     }
@@ -20,18 +23,14 @@ public class DriveMotorsSpeedTest extends CommandBase {
     @Override
     public void initialize() {
         timer.start();
+        Arrays.fill(states, new SwerveModuleState(0, Rotation2d.fromDegrees(90)));
     }
 
     @Override
     public void execute() {
-        swerveDrive.setStates(new SwerveModuleState[]{
-                new SwerveModuleState(3, Rotation2d.fromDegrees(0)),
-                new SwerveModuleState(3, Rotation2d.fromDegrees(0)),
-                new SwerveModuleState(3, Rotation2d.fromDegrees(0)),
-                new SwerveModuleState(3, Rotation2d.fromDegrees(0))
-        });
+        swerveDrive.setStates(states);
         for (int i = 0; i < 4; i++) {
-            if (Math.abs(Math.abs(swerveDrive.getModule(i).getVelocity()) - 3) < 0.1) {
+            if (Math.abs(swerveDrive.getModule(i).getAngle().minus(Rotation2d.fromDegrees(90)).getDegrees()) < 3) {
                 if (times[i] == 0) times[i] = timer.get();
             }
         }
@@ -40,8 +39,8 @@ public class DriveMotorsSpeedTest extends CommandBase {
     @Override
     public void end(boolean interrupted) {
         for (int i = 0; i < 4; i++) {
-            System.out.println("WHEEL: " + i + " reached 3m/s in: " + times[i] + " seconds.");
-            SmartDashboard.putString("SpeedTestResult [WHEEL " + i + "]", "WHEEL: " + i + " reached 3m/s in: " + times[i] + " seconds.");
+            System.out.println("WHEEL: " + i + " passed 90 degrees in: " + times[i] + " seconds.");
+            SmartDashboard.putString("SpeedTestResult [WHEEL " + i + "]", "WHEEL: " + i + " passed 90 degrees in: " + times[i] + " seconds.");
         }
         swerveDrive.terminate();
     }
