+++ src/main/java/frc/robot/subsystems/drivetrain/commands/tests/DriveMotorsEqualVelocityTest.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.drivetrain.tests;
+package frc.robot.subsystems.drivetrain.commands.tests;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
