+++ src/main/java/frc/robot/subsystems/drivetrain/commands/tests/DriveMotorsSpeedTest.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.drivetrain.tests;
+package frc.robot.subsystems.drivetrain.commands.tests;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
@@ -7,15 +7,12 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.drivetrain.SwerveDrive;
 
-import java.util.Arrays;
-
-public class AngleMotorsSpeedTest extends CommandBase {
+public class DriveMotorsSpeedTest extends CommandBase {
     private final SwerveDrive swerveDrive;
     private final Timer timer = new Timer();
     private final double[] times = new double[4];
-    private final SwerveModuleState[] states = new SwerveModuleState[4];
 
-    public AngleMotorsSpeedTest(SwerveDrive swerveDrive) {
+    public DriveMotorsSpeedTest(SwerveDrive swerveDrive) {
         this.swerveDrive = swerveDrive;
         addRequirements(swerveDrive);
     }
@@ -23,14 +20,18 @@ public class AngleMotorsSpeedTest extends CommandBase {
     @Override
     public void initialize() {
         timer.start();
-        Arrays.fill(states, new SwerveModuleState(0, Rotation2d.fromDegrees(90)));
     }
 
     @Override
     public void execute() {
-        swerveDrive.setStates(states);
+        swerveDrive.setStates(new SwerveModuleState[]{
+                new SwerveModuleState(3, Rotation2d.fromDegrees(0)),
+                new SwerveModuleState(3, Rotation2d.fromDegrees(0)),
+                new SwerveModuleState(3, Rotation2d.fromDegrees(0)),
+                new SwerveModuleState(3, Rotation2d.fromDegrees(0))
+        });
         for (int i = 0; i < 4; i++) {
-            if (Math.abs(swerveDrive.getModule(i).getAngle().minus(Rotation2d.fromDegrees(90)).getDegrees()) < 3) {
+            if (Math.abs(Math.abs(swerveDrive.getModule(i).getVelocity()) - 3) < 0.1) {
                 if (times[i] == 0) times[i] = timer.get();
             }
         }
@@ -39,8 +40,8 @@ public class AngleMotorsSpeedTest extends CommandBase {
     @Override
     public void end(boolean interrupted) {
         for (int i = 0; i < 4; i++) {
-            System.out.println("WHEEL: " + i + " passed 90 degrees in: " + times[i] + " seconds.");
-            SmartDashboard.putString("SpeedTestResult [WHEEL " + i + "]", "WHEEL: " + i + " passed 90 degrees in: " + times[i] + " seconds.");
+            System.out.println("WHEEL: " + i + " reached 3m/s in: " + times[i] + " seconds.");
+            SmartDashboard.putString("SpeedTestResult [WHEEL " + i + "]", "WHEEL: " + i + " reached 3m/s in: " + times[i] + " seconds.");
         }
         swerveDrive.terminate();
     }
